Adds check for clobbering of files
When two users have the same Google Drive file open (or even one user has the same file open in multiple tabs) these files can clobber each other.  This makes for a very bad user experience.  This adds a warning that the file has been modified by another user since it was opened in this tab, and gives the user the option to save anyay (overwriting the work of other users) or cancel.