Use Google FilePicker to open files
When particular file was not created with this app, or has not been opened in the FilePicker before with this app, the user cannot load that file.  This is a security measure that prevents our app accessing arbitrary user data (see drive.file scope in Google API documentation).  This commit prompts the user to select a file in the FilePicker, whenever we need access to a file.