Updates clobber detection and removes checkpointing
The new behavior of Google Drive requires a new system for detecting clobbering by other users, and also breaks checkpointing.  This commit adds a new method for detecting clobbering, by writing a custom property to the files metadata with each save.  If an unexpected value of this property is observed, this indicates that another user is modifying this file.  This method is not failsafe, but shoud be reliable if users dont save to frequently.  Checkpointing is removed because there is no way to reliably checkpoint when we don't know what the last saved revision is.