.gitignore: add .*.swo pattern
As noted in 5e427d990fe, a .swo file was inadvertently added to the repo. Add a
line to the .gitignore file to stop this happening in future.

(.swo files are created by vim if a) a .swp file already exists for a file and
b) it is opened by another editor.)