gapiutils: defensively re-auth on 401 status
In Google's FAQ[1], the preferred option for dealing with auth token expiry is
to refresh it before it expires. The timeout may not fire if, for example, the
laptop with the notebook page displayed is put to sleep.

Add a defensive measure to the wrapped execute() call which will attempt
re-authorization when a 401 is received. The reauthorization is only attempted
once. If a 401 is received again, the failure is passed back to the caller.

[1] https://developers.google.com/api-client-library/javascript/help/faq